# backend/internal/adapters/http/handlers/openapi/swagger.yaml
components:
  schemas:
...
    dto.CreateLabelRequest:
      properties:
        color:
          type: string
        created_at:
          format: date-time
          type: string
        id:
          type: string
        name:
          type: string
        owner:
          enum:
          - personal
          - family
          - system
          example: personal
          type: string
      required:
      - color
      - name
      - owner
      type: object
...
    dto.LabelModel:
      properties:
        color:
          description: '@Description Hexadecimal color code for visual representation
            of the label'
          example: '#FF5733'
          pattern: ^#[0-9A-Fa-f]{6}$
          type: string
        created_at:
          description: '@Description ISO 8601 timestamp indicating when the label
            was originally created'
          example: "2023-01-15T10:30:00Z"
          format: date-time
          type: string
        etag:
          description: '@Description Entity tag used for optimistic concurrency control
            to prevent conflicting updates'
          example: W/"123456789"
          type: string
        id:
          description: '@Description Unique identifier for the label (UUID format)'
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        key:
          type: string
        name:
          description: '@Description Display name of the label'
          example: Entertainment
          maxLength: 100
          minLength: 1
          type: string
        owner:
          $ref: '#/components/schemas/dto.OwnerModel'
        updated_at:
          description: '@Description ISO 8601 timestamp indicating when the label
            was last modified'
          example: "2023-01-20T14:45:30Z"
          format: date-time
          type: string
      required:
      - color
      - created_at
      - etag
      - id
      - name
      - owner
      - updated_at
      type: object
...
    dto.OwnerModel:
      description: '@Description Ownership information specifying whether this subscription
        belongs to a user or family'
      properties:
        etag:
          description: '@Description Entity tag for optimistic concurrency control'
          example: W/"123456789"
          type: string
        family_id:
          description: '@Description Family LabelID when an ownership type is family
            (required for family ownership)'
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        type:
          description: '@Description Type of ownership (personal, family or system)'
          enum:
          - personal
          - family
          - system
          example: personal
          type: string
        userId:
          description: '@Description UserProfile LabelID when an ownership type is
            personal (required for personal ownership)'
          example: 123e4567-e89b-12d3-a456-426614174001
          type: string
      required:
      - etag
      - type
      type: object
...
    dto.PaginatedResponseModel-dto_LabelModel:
      properties:
        data:
          description: Data contains the list of items for the current page
          items:
            $ref: '#/components/schemas/dto.LabelModel'
          type: array
          uniqueItems: false
        length:
          description: Length represents the number of items in the current page
          type: integer
        total:
          description: Total represents the total number of items available
          type: integer
      required:
      - data
      - length
      - total
      type: object
    dto.ProviderModel:
      description: Provider object containing information about a subscription service
        provider and their available plans
      properties:
        created_at:
          description: '@Description ISO 8601 timestamp when the provider was originally
            created'
          example: "2023-01-15T10:30:00Z"
          format: date-time
          type: string
        description:
          description: '@Description Optional detailed description of the provider
            and their services'
          example: Streaming service offering movies and TV shows
          type: string
        etag:
          description: '@Description Entity tag used for optimistic concurrency control
            to prevent conflicting updates'
          example: W/"123456789"
          type: string
        icon_url:
          description: '@Description Optional URL to the provider''s icon or logo
            image'
          example: https://example.com/netflix-icon.png
          type: string
        id:
          description: '@Description Unique identifier for the provider (UUID format)'
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        key:
          example: netflix
          maxLength: 255
          minLength: 1
          type: string
        labels:
          description: '@Description List of label IDs associated with this provider
            for categorization'
          example:
          - 123e4567-e89b-12d3-a456-426614174001
          - 123e4567-e89b-12d3-a456-426614174002
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          description: '@Description Display name of the service provider'
          example: Netflix
          maxLength: 255
          minLength: 1
          type: string
        owner:
          $ref: '#/components/schemas/dto.OwnerModel'
        pricing_page_url:
          description: '@Description Optional URL to the provider''s pricing information
            page'
          example: https://netflix.com/pricing
          type: string
        updated_at:
          description: '@Description ISO 8601 timestamp when the provider was last
            modified'
          example: "2023-01-20T14:45:30Z"
          format: date-time
          type: string
        url:
          description: '@Description Optional URL to the provider''s main website'
          example: https://netflix.com
          type: string
      required:
      - created_at
      - etag
      - id
      - key
      - labels
      - name
      - owner
      - updated_at
      type: object
...
    dto.UpdateProviderRequest:
      properties:
        description:
          type: string
        icon_url:
          type: string
        labels:
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          type: string
        pricing_page_url:
          type: string
        updated_at:
          format: date-time
          type: string
        url:
          type: string
      required:
      - labels
      - name
      type: object
...
    ginx.HttpErrorResponse:
      description: RFC7807 Problem Details error response
      properties:
        detail:
          example: Missing required field 'name'
          type: string
        instance:
          example: /api/resource/123
          type: string
        status:
          example: 400
          type: integer
        title:
          example: Bad Request
          type: string
        type:
          example: about:blank
          type: string
      type: object
...
info:
  contact:
    email: support@mistribe.com
    name: API Support
    url: http://subtracker.mistribe.com/support
  description: This api provide HTTPRest endpoints for the application SubTracker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://subtracker.mistribe.com/terms/
  title: SubTracker API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /labels:
    get:
      description: Retrieve a paginated list of labels with optional filtering by
        owner type and search text
      parameters:
      - description: Search text to filter labels by name
        in: query
        name: search
        schema:
          type: string
      - description: 'Maximum number of items to return (default: 10)'
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Number of items to skip for pagination (default: 0)'
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.PaginatedResponseModel-dto_LabelModel'
          description: Paginated list of labels
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Get all labels
      tags:
      - labels
    post:
      description: Create a new label with specified name, color, and owner information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.CreateLabelRequest'
        description: Label creation data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.LabelModel'
          description: Successfully created label
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid input data
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Unauthorized - Invalid user authentication
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Create a new label
      tags:
      - labels
...
  /labels/export:
    get:
      description: Export all labels in CSV, JSON, or YAML format
      parameters:
      - description: Export format (csv, json, yaml)
        in: query
        name: format
        schema:
          default: json
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/x-yaml:
              schema:
                type: string
          description: Exported labels file
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Invalid format parameter
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Export labels
      tags:
      - labels
...
  /providers/{providerId}:
    delete:
      description: Permanently delete a provider and all its associated plans and
        prices
      parameters:
      - description: Provider LabelID (UUID format)
        in: path
        name: providerId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content - Provider successfully deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid provider LabelID format
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Provider not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Delete provider by LabelID
      tags:
      - providers
    get:
      description: Retrieve a single provider with all its plans and prices by LabelID
      parameters:
      - description: Provider ID (UUID format) or Provider Key (string format)
        in: path
        name: providerId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ProviderModel'
          description: Successfully retrieved provider
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid provider LabelID format
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Provider not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Get provider by LabelID
      tags:
      - providers
    put:
      description: Update an existing provider's basic information
      parameters:
      - description: Provider LabelID (UUID format)
        in: path
        name: providerId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.UpdateProviderRequest'
        description: Updated provider data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ProviderModel'
          description: Successfully updated provider
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid input data or provider LabelID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Provider not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Update provider by LabelID
      tags:
      - providers
  /providers/export:
    get:
      description: Export all providers in CSV, JSON, or YAML format
      parameters:
      - description: Export format (csv, json, yaml)
        in: query
        name: format
        schema:
          default: json
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/x-yaml:
              schema:
                type: string
          description: Exported providers file
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Invalid format parameter
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Export providers
      tags:
      - providers
...
servers:
- description: Production server
  url: https://api.subtracker.mistribe.com
- description: Development server
  url: http://localhost:8080